refactor: ensure that conversationindexconstants reference mlindex enums rather than use their own mappings
